创建keyboard数据模型

+++ app/src/main/java/sunhang/mathkeyboard/kbdinfo/InfosFactory.kt
@@ -8,5 +8,5 @@ abstract class InfosFactory(protected val context: Context) {
     protected val keyHorPadding: Float = context.resources.getDimensionPixelSize(R.dimen.key_horizontal_padding).toFloat()
     protected val keyVerPadding: Float = context.resources.getDimensionPixelSize(R.dimen.key_vertical_padding).toFloat()
 
-    abstract fun createKeyboardInfo(keyboardWidth: Float, keyboardHeight: Float): KbdInfo.KeyboardInfo
+    abstract fun createKeyboardInfo(keyboardWidth: Int, keyboardHeight: Int): KbdInfo.KeyboardInfo
 }
